Update daemon.sh
adding to crontab 
*/2 * * * * ( /Data/bypy/daemon.sh > /dev/null 2>&1 )

+++ daemon.sh
@@ -6,13 +6,21 @@ pythonExec=/usr/local/bin/python2.7 # python binary (version > 2.7)
 
 if  [[ `ps -elf|grep bypy|grep syncup|wc -l` -eq 0 ]]; then  
     echo `date` > $logFile
+    cd $dataDir
+    #recording the current file numbers
+    lines_bef=`ls -l|wc -l`
     $pythonExec $execDir/bypy.py syncup $dataDir >> $logFile 2>&1
     if [[ $? -eq 0 ]]; then
-        echo "syncup down, clean and recycle local dir" >> $logFile
-        rm -fr $dataDir/*
+        lines_cur=`ls -l|wc -l`
+        if [ $lines_bef -eq $lines_cur ]; then 
+            echo "syncup down, clean and recycle local dir" >> $logFile
+            rm -fr $dataDir/*
+        else
+            echo "$(date) adding more files during sync time, so don't clean $dataDir right now, delay to next time"
+        fi
     else
-        echo "Err: return status: $?" >> $logFile
+        echo "$(date) Err: return status: $?" >> $logFile
     fi
 else
-    echo "bypy currently is running $(date)!" >> $logFile
+    echo "$(date) bypy currently is running !" >> $logFile
 fi
